+++ packages/caelus/src/index.ts
@@ -24,3 +24,4 @@ export * from "./directions.js";
 export * from "./vargas.js";
 export * from "./yogini.js";
 export * from "./yogas.js";
+export * from "./ashtottari.js";